Big Gulp 4 update
- Update to Gulp 4 and updated the gulpfile to the new API
- Update all other dependencies to latest versions to address security
  alerts.
- No longer using Babel to transpile the gulpfile.
