+++ package.json
@@ -3,7 +3,7 @@
   "private": true,
   "version": "0.0.1",
   "description": "Sphinx theme for LSST docs, based on Bootstrap and Astropy.",
-  "main": "gulpfile.babel.js",
+  "main": "gulpfile.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "precommit": "pretty-quick --staged",
@@ -19,24 +19,19 @@
     "url": "https://github.com/lsst-sqre/lsst-sphinx-bootstrap-theme/issues"
   },
   "homepage": "https://github.com/lsst-sqre/lsst-sphinx-bootstrap-theme#readme",
-  "dependencies": {
-    "gulp-run": "^1.7.1"
-  },
+  "dependencies": {},
   "devDependencies": {
-    "autoprefixer": "^8.5.0",
-    "babel": "^6.23.0",
-    "babel-cli": "^6.26.0",
-    "babel-core": "^6.26.3",
-    "babel-preset-env": "^1.7.0",
-    "babel-register": "^6.26.0",
-    "gulp": "^3.9.1",
-    "gulp-clean-css": "^3.9.4",
-    "gulp-postcss": "^7.0.1",
-    "gulp-sass": "^4.0.1",
-    "gulp-sourcemaps": "^2.6.4",
-    "husky": "^0.14.3",
+    "autoprefixer": "^9.6.1",
+    "gulp": "^4.0.2",
+    "gulp-clean-css": "^4.2.0",
+    "gulp-noop": "^1.0.0",
+    "gulp-postcss": "^8.0.0",
+    "gulp-run": "^1.7.1",
+    "gulp-sass": "^4.0.2",
+    "gulp-sourcemaps": "^2.6.5",
+    "husky": "^3.0.2",
     "minimist": "^1.2.0",
-    "prettier": "1.12.1",
-    "pretty-quick": "^1.4.1"
+    "prettier": "1.18.2",
+    "pretty-quick": "^1.11.1"
   }
 }
